New outputs. Referring to first commit of this push:
Modified several files will go through one by one, each getting their own header.

Added the `pymongo` dependency to the depedencies list.

Configured logging to go to the new `src/academic_metrics/logs/` directory.

Same with `ChainBuilder.py`, configured the logging to go to the new `logs/` directory.

Moved the find directory utilility functions here to allow for importing of the constant `LOG_FILE_DIR` to all files without circular import issues.

Added the `traceback` python module to implement more thorough error logging to help with debugging edge cases introduced by the fetching of missing abstracts from the use of selenium and an AI chain. I think the route to go here to avoid edge cases is to add the extra_context variable returned by the chain in `scraper.py` to the return back to `CrossrefWrapper.py` and it can just be dumped into the classification prompt, that way if ther is an abstract missing we don't have to worry about it's item actually getting filtered out.

I think this happens as the chain in `scraper.py` instructs the LLM to still provide the abstract key even if it can't find it and that if it can't find it to set it as an empty string. The empty string is checked for and `scraper.py` returns `None` if the abstract wasn't found, then when `CrossrefWrapper.py` sees that the abstract is none it removes the item. Well this works, but the issue is the LLM used in the chain from `scraper.py` sometimes returns a single space within the abstract field. This inconsitency is going to be hard to plan for, so there's 2 options to solve it.

Add a 3rd field to the return statement from the LLM. That field can be a boolean `found_abstract` that way the AI sets it to one of the booleans, then instead of checking if abstract exists we check if `found_abstract==True`.

The second option is to just not care what the abstract is, and always return out `extra_context` then push that `extra_context` map into the classification prompt and adjusting the prompt with something like:

```python
"""You will be provided an abstract... you will also be provided some extra_context... use the abstract and the extra_context to classify the paper... sometimes there may not be an abstract available, in that case use the extra context to do the classification