Changes to avoid hanging forever due to invalid classification
After moving to month processing, and fixing all other hanging issues, I still noticed hanging once the pipeline hit CategoryProcessor. I discovered it was when generating the url. Further investingation I discovered it was when making a call to Taxonomy's method get_mid_cat_for_low_cat(), an invalid category that the AI provided would get passed here and the loop which looked to find the mid category for the low category would never complete as it was a low category which did not exist.

Detailed below is a breakown of the changes to each file which solves this issue.

taxonomy_util.py:
Added 3 instance varialbes all_top_categories, all_mid_categories, and all_low_categories, and setter methods which will initialize them.

Added a method 'is_valid_category' which takes in the category name and a level and returns a boolean indicating category validity.

ChainBuilder.py:

Added a param 'words_to_ban' which is to be a list of strings for words to ban out of the AI generation.

This is done by using the ChatOpenAI param logit_bias like so:

logit_bias biases the generation of tokens. It is of type Dict[int, int] where the key is the token, and the integer is the bias. To ban the words the bias is set to -100.

To allow for dynamically setting this a setter method 'set_words_to_ban' was added, whcih takes in the list, validates it, get the logit_bias for it, and then reinitializes the ChatOpenAI object with the new logit_bias.

The logit_bias is calculated using tiktoken in the method '_get_logit_bias_dict'.

AbstractClassifier.py:

Added an instance variable 'self.words_to_ban'

In the method classify_abstract the following logic was created:

Upon a classification run finishing and the categories being extracted, a while loop with condition 'while any categories in classified categories are not valid' where it will find the invalid categories using the 'is_valid_category' method which wraps around Taxonomy's 'is_valid_category' method. It will then check to ensure we haven't tried more than the instance variable 'self.max_classification_retries' (if we have it will raise a value error which will be caught by classification_orchestrator resulting in the item being popped out of the data).

If we aren't on the last retry we will simply try again.

If we are on the last retry we will use the self.banned_words instance variable we update with the discovered invalid categories and then set the words_to_ban instance variable in ChainManager via the set_words_to_ban method. Then we will try again with those words banned.

The reason this is only done on the last try is because it's possible a word is more than 1 token. If it is more than one token then that is going to make it to where the LLM will never generate that token, if that token is part of a category it would've been classified into then the LLM is not likely to generate that category and we don't get it when we otherwise would. Since it's better to have some classifications rather than just not have the item of data at all, we do it on the last try.

CrossrefWrapper.py: Set up to be able to run on individual months.

pipeline.py: set up to be able to configure CrossrefWrapper on individual months.

Also set it to run from 2017-2022 one month at a time, that way if we hit an error in the 6th month or 8th month or 3rd month, we would've already pushed all of the data for months before that. This means we get more data quickly, and avoid wasting API costs.

CrossrefWrapper.py: method added to ensure we make the correct last day depdending on the month.

Other changes are mostly adding logging where there wasn't any.