Edge cases handled and all data from 2017-2024 processed. Changes detailed below
AbstractClassifier.py: Fixed referencing issue of 'self.max_retries', it was supposed to be 'self.max_classification_retries' and now is.

Also made the models configurable.

abstract_classifier_factory:

function takes in the models for which to use.

classification_orchestrator:
the run method takes in the models for which to use.

pipeline.py:

added models as params

also added ability to add in the models as arguments parsing like so:
python pipeline.py --pre-classification-model="gpt-4o-min" --classification-model="gpt-4o" --theme-model="gpt-4o-mini"