System updates and documentation updates
docs:

Overhaul of documentation to use a different theme and several optional params to make it look and feel better as well as make it easier to navigate.

constants package added to docs

rst files updated.

AbstractClassifier.py renamed to abstract_classifier.py to make it more clear when importing which is the file and which is the class.

Updates to all docs strings.

Postprocessing:

FacultyPostprocessor split up.

Base class implemented which takes in an attribute name rather than be hard coded for faculty.

2 child classes for faculty and departments made and tested. They act as simple wrappers, technically they're redudant code, but I do think it makes the code more readable when it's being used.

Post processing moved out of core/ and into it's own postprocessing/ submodule.

Excel script and file generated.